roscpp: TransportTCP: replace recursive_mutex by mutex
This makes everything much simpler. We have a small set of private
functions that are externally locked (i.e. assume the mutex is acquired),
and a set of public functions, which acquire the mutex when required.
Calling a public function while holding the mutex is *not* allowed!

With the old design it was very much possible to call close() with the
mutex held, leading to deadlock hell...